remove DirectClientV1
